+++ ApplicationLibCode/Application/CMakeLists_files.cmake
@@ -38,6 +38,7 @@ set(SOURCE_GROUP_HEADER_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RiaOsduDefines.h
     ${CMAKE_CURRENT_LIST_DIR}/RiaCloudDefines.h
     ${CMAKE_CURRENT_LIST_DIR}/RiaWellFlowDefines.h
+    ${CMAKE_CURRENT_LIST_DIR}/RiaWellLogTrackDefines.h
     ${CMAKE_CURRENT_LIST_DIR}/RiaQuickAccessScheduler.h
     ${CMAKE_CURRENT_LIST_DIR}/RiaGridDefines.h
     ${CMAKE_CURRENT_LIST_DIR}/RiaModelExportDefines.h
@@ -83,6 +84,7 @@ set(SOURCE_GROUP_SOURCE_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RiaOsduDefines.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RiaCloudDefines.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RiaWellFlowDefines.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/RiaWellLogTrackDefines.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RiaQuickAccessScheduler.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RiaGridDefines.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RiaModelExportDefines.cpp
