+++ ApplicationLibCode/Application/RiaWellLogTrackDefines.cpp
@@ -16,7 +16,7 @@
 //
 /////////////////////////////////////////////////////////////////////////////////
 
-#include "RimWellLogTrackEnums.h"
+#include "RiaWellLogTrackDefines.h"
 
 namespace caf
 {
