+++ ApplicationLibCode/Commands/RicWellLogTools.cpp
@@ -38,7 +38,6 @@
 #include "RimWellLogLasFileCurve.h"
 #include "RimWellLogRftCurve.h"
 #include "RimWellLogTrack.h"
-#include "RimWellLogTrackEnums.h"
 #include "RimWellLogWbsCurve.h"
 #include "RimWellMeasurementCurve.h"
 #include "RimWellPath.h"
