+++ ApplicationLibCode/ProjectDataModel/WellLog/CMakeLists_files.cmake
@@ -7,7 +7,6 @@ set(SOURCE_GROUP_HEADER_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RimWellLogFormationSettings.h
     ${CMAKE_CURRENT_LIST_DIR}/RimWellLogRegionAnnotationSettings.h
     ${CMAKE_CURRENT_LIST_DIR}/RimWellLogWellPathAttributeSettings.h
-    ${CMAKE_CURRENT_LIST_DIR}/RimWellLogTrackEnums.h
     ${CMAKE_CURRENT_LIST_DIR}/RimWellLogPlotNameConfig.h
     ${CMAKE_CURRENT_LIST_DIR}/RimWellLogExtractionCurveNameConfig.h
     ${CMAKE_CURRENT_LIST_DIR}/RimWellLogFileCurveNameConfig.h
@@ -50,7 +49,6 @@ set(SOURCE_GROUP_SOURCE_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RimWellLogFormationSettings.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RimWellLogRegionAnnotationSettings.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RimWellLogWellPathAttributeSettings.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/RimWellLogTrackEnums.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RimWellLogTrack.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RimWellLogCurve.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RimWellLogExtractionCurve.cpp
