+++ ApplicationLibCode/ProjectDataModel/WellLog/RimWellLogCurveCommonDataSource.cpp
@@ -42,7 +42,6 @@
 #include "RimWellLogPlotCollection.h"
 #include "RimWellLogRftCurve.h"
 #include "RimWellLogTrack.h"
-#include "RimWellLogTrackEnums.h"
 #include "RimWellLogWbsCurve.h"
 #include "RimWellMeasurementCurve.h"
 #include "RimWellPath.h"
